# lib/scanner: Simplify, optimize and document Validate (#6674)